+++ app/src/main/java/com/example/playcoin/domain/model/Coin.kt
@@ -4,7 +4,7 @@ data class Coin(
     val id: String,
     val name: String,
     val image: String,
-    val market_cap: Long,
+    val market_cap: Double,
     val price: Double,
     val price_percentage_change: Double,
     val low_price: Double,
